requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
anyio==4.12.0
asgiref==3.11.0
attrs==25.4.0
bidict==0.23.1
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cssselect2==0.8.0
Django==5.2.9
docutils==0.22.3
fastapi==0.124.0
filelock==3.20.0
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.12.0
h11==0.16.0
htmx==0.0.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown2==2.5.4
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
networkx==3.6
nicegui==3.3.1
numpy==2.3.5
orjson==3.11.5
packaging==25.0
pillow==12.0.0
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==1.10.24
pydyf==0.12.1
Pygments==2.19.2
pyphen==0.17.2
python-dotenv==1.2.1
python-engineio==4.12.3
python-multipart==0.0.20
python-socketio==5.15.0
pypdf==5.1.0
PyYAML==6.0.3
qrcode==8.2
regex==2025.11.3
safetensors==0.7.0
setuptools==80.9.0
simple-websocket==1.1.0
sqlparse==0.5.4
starlette==0.50.0
sympy==1.14.0
tinycss2==1.5.1
tinyhtml5==2.0.0
torch==2.9.1
transformers==4.57.3
llama-cpp-python
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
uvicorn==0.38.0
watchfiles==1.1.1
weasyprint==67.0
webencodings==0.5.1
websockets==15.0.1
wsproto==1.3.2
yarl==1.22.0
zopfli==0.4.0
